Fix refresh token race condition
