+++ frontend/src/App.tsx
@@ -7,7 +7,7 @@ import * as Sentry from "@sentry/react";
 import { QueryCache, QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import axios from "axios";
 import { useEffect, useMemo } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import {
 	RouterProvider,
 	createBrowserRouter,
@@ -21,10 +21,7 @@ import { useGetSettingsQuery } from "./hooks/api/settings";
 import { useErrorHandler } from "./hooks/useErrorHandler";
 import AuthProvider from "./providers/AuthProvider";
 import { routes } from "./routes";
-import { logout, setAccessToken } from "./store/authSlice";
 import { selectConfig } from "./store/configSlice";
-import { store } from "./store/store";
-import { setUserData } from "./store/userSlice";
 import { isAxios404Error, isAxios500Error } from "./utils/axios.default";
 
 if (import.meta.env.PROD) {
@@ -61,7 +58,6 @@ function SettingLoader() {
 
 function App() {
 	const config = useSelector(selectConfig);
-	const dispatch = useDispatch();
 	const prefersDarkMode = useMediaQuery("(prefers-color-scheme: dark)");
 	const theme = useMemo(() => {
 		const defaultMode = prefersDarkMode ? "dark" : "light";
@@ -107,39 +103,6 @@ function App() {
 			},
 		});
 	}, [handleError]);
-
-	useEffect(() => {
-		const handleRefreshTokenExpiration = () => {
-			dispatch(logout());
-			dispatch(setUserData(null));
-		};
-
-		const interceptor = axios.interceptors.response.use(
-			(response) => response,
-			async (error) => {
-				if (error.response?.status === 401 && !error.config._retry) {
-					if (error.config.url === "/auth/refresh") {
-						handleRefreshTokenExpiration();
-						return Promise.reject(error);
-					} else {
-						error.config._retry = true;
-						const refreshToken = store.getState().auth.refreshToken;
-						const response = await axios.post("/auth/refresh", { refreshToken });
-						const newAccessToken = response.data.newAccessToken;
-						dispatch(setAccessToken(newAccessToken));
-						axios.defaults.headers.common["Authorization"] = `Bearer ${newAccessToken}`;
-						error.config.headers["Authorization"] = `Bearer ${newAccessToken}`;
-						return axios(error.config);
-					}
-				}
-				return Promise.reject(error);
-			}
-		);
-
-		return () => {
-			axios.interceptors.response.eject(interceptor);
-		};
-	}, [dispatch]);
 
 	return (
 		<QueryClientProvider client={queryClient}>
